+++ src/devices/evse.test.ts
@@ -19,11 +19,11 @@ import { Identify, PowerSource, ElectricalEnergyMeasurement, ElectricalPowerMeas
 import { EnergyEvseServer, EnergyEvseModeServer, DeviceEnergyManagementModeServer } from '@matter/node/behaviors';
 
 // Matterbridge
-import { MatterbridgeEndpoint } from '../matterbridgeEndpoint.ts';
-import { invokeBehaviorCommand } from '../matterbridgeEndpointHelpers.ts';
-import { MatterbridgeDeviceEnergyManagementModeServer } from '../matterbridgeBehaviors.ts';
+import { MatterbridgeEndpoint } from '../matterbridgeEndpoint.js';
+import { invokeBehaviorCommand } from '../matterbridgeEndpointHelpers.js';
+import { MatterbridgeDeviceEnergyManagementModeServer } from '../matterbridgeBehaviors.js';
 
-import { Evse, MatterbridgeEnergyEvseServer, MatterbridgeEnergyEvseModeServer } from './evse.ts';
+import { Evse, MatterbridgeEnergyEvseServer, MatterbridgeEnergyEvseModeServer } from './evse.js';
 
 let loggerLogSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.log>;
 let consoleLogSpy: jest.SpiedFunction<typeof console.log>;
