+++ src/devices/laundryWasher.ts
@@ -26,9 +26,7 @@ import { MaybePromise } from '@matter/main';
 import { OperationalState } from '@matter/main/clusters/operational-state';
 import { LaundryWasherControls } from '@matter/main/clusters/laundry-washer-controls';
 import { LaundryWasherMode } from '@matter/main/clusters/laundry-washer-mode';
-import { TemperatureControl } from '@matter/main/clusters/temperature-control';
 import { ModeBase } from '@matter/main/clusters/mode-base';
-import { TemperatureControlServer } from '@matter/main/behaviors/temperature-control';
 import { LaundryWasherModeServer } from '@matter/main/behaviors/laundry-washer-mode';
 import { LaundryWasherControlsServer } from '@matter/main/behaviors/laundry-washer-controls';
 
@@ -36,6 +34,8 @@ import { LaundryWasherControlsServer } from '@matter/main/behaviors/laundry-wash
 import { laundryWasher, powerSource } from '../matterbridgeDeviceTypes.js';
 import { MatterbridgeEndpoint } from '../matterbridgeEndpoint.js';
 import { MatterbridgeOnOffServer, MatterbridgeServer } from '../matterbridgeBehaviors.js';
+
+import { createLevelTemperatureControlClusterServer, createNumberTemperatureControlClusterServer } from './temperatureControl.js';
 
 export class LaundryWasher extends MatterbridgeEndpoint {
   /**
@@ -85,8 +85,8 @@ export class LaundryWasher extends MatterbridgeEndpoint {
     this.createDeadFrontOnOffClusterServer(true);
     this.createDefaultLaundryWasherModeClusterServer(currentMode, supportedModes);
     this.createDefaultLaundryWasherControlsClusterServer(spinSpeedCurrent, spinSpeeds, numberOfRinses, supportedRinses);
-    if (temperatureSetpoint) this.createNumberTemperatureControlClusterServer(temperatureSetpoint, minTemperature, maxTemperature, step);
-    else this.createLevelTemperatureControlClusterServer(selectedTemperatureLevel, supportedTemperatureLevels);
+    if (temperatureSetpoint) createNumberTemperatureControlClusterServer(this, temperatureSetpoint, minTemperature, maxTemperature, step);
+    else createLevelTemperatureControlClusterServer(this, selectedTemperatureLevel, supportedTemperatureLevels);
     this.createDefaultOperationalStateClusterServer(operationalState);
   }
 
@@ -137,81 +137,6 @@ export class LaundryWasher extends MatterbridgeEndpoint {
       numberOfRinses, // Writable
     });
     return this;
-  }
-
-  /**
-   * Creates a TemperatureControl Cluster Server with feature TemperatureLevel.
-   *
-   * @param {number} selectedTemperatureLevel - The selected temperature level as an index of the supportedTemperatureLevels array. Defaults to 1 (which corresponds to 'Warm').
-   * @param {string[]} supportedTemperatureLevels - The supported temperature levels. Defaults to ['Cold', 'Warm', 'Hot', '30°', '40°', '60°', '80°'].
-   *
-   * @returns {this} The current MatterbridgeEndpoint instance for chaining.
-   */
-  createLevelTemperatureControlClusterServer(selectedTemperatureLevel: number = 1, supportedTemperatureLevels: string[] = ['Cold', 'Warm', 'Hot', '30°', '40°', '60°', '80°']): this {
-    this.behaviors.require(MatterbridgeLevelTemperatureControlServer.with(TemperatureControl.Feature.TemperatureLevel), {
-      selectedTemperatureLevel,
-      supportedTemperatureLevels,
-    });
-    return this;
-  }
-  /**
-   * Creates a TemperatureControl Cluster Server with features TemperatureNumber and TemperatureStep.
-   *
-   * @param {number} temperatureSetpoint - The temperature setpoint * 100. Defaults to 40 * 100 (which corresponds to 40°C).
-   * @param {number} minTemperature - The minimum temperature * 100. Defaults to 30 * 100 (which corresponds to 30°C). Fixed attribute.
-   * @param {number} maxTemperature - The maximum temperature * 100. Defaults to 60 * 100 (which corresponds to 60°C). Fixed attribute.
-   * @param {number} [step] - The step size for temperature changes. Defaults to 10 * 100 (which corresponds to 10°C). Fixed attribute.
-   *
-   * @returns {this} The current MatterbridgeEndpoint instance for chaining.
-   */
-  createNumberTemperatureControlClusterServer(temperatureSetpoint: number = 40 * 100, minTemperature: number = 30 * 100, maxTemperature: number = 60 * 100, step: number = 10 * 100): this {
-    this.behaviors.require(MatterbridgeNumberTemperatureControlServer.with(TemperatureControl.Feature.TemperatureNumber, TemperatureControl.Feature.TemperatureStep), {
-      temperatureSetpoint,
-      minTemperature, // Fixed attribute
-      maxTemperature, // Fixed attribute
-      step, // Fixed attribute
-    });
-    return this;
-  }
-}
-
-export class MatterbridgeLevelTemperatureControlServer extends TemperatureControlServer.with(TemperatureControl.Feature.TemperatureLevel) {
-  override initialize() {
-    if (this.state.supportedTemperatureLevels.length >= 2) {
-      const device = this.endpoint.stateOf(MatterbridgeServer);
-      device.log.info(`MatterbridgeLevelTemperatureControlServer initialized with selectedTemperatureLevel ${this.state.selectedTemperatureLevel} and supportedTemperatureLevels: ${this.state.supportedTemperatureLevels.join(', ')}`);
-    }
-  }
-
-  override setTemperature(request: TemperatureControl.SetTemperatureRequest): MaybePromise {
-    const device = this.endpoint.stateOf(MatterbridgeServer);
-    device.log.info(`SetTemperature (endpoint ${this.endpoint.maybeId}.${this.endpoint.maybeNumber})`);
-    device.commandHandler.executeHandler('setTemperature', { request, cluster: TemperatureControlServer.id, attributes: this.state, endpoint: this.endpoint });
-    if (request.targetTemperatureLevel !== undefined && request.targetTemperatureLevel >= 0 && request.targetTemperatureLevel < this.state.supportedTemperatureLevels.length) {
-      device.log.debug(`MatterbridgeLevelTemperatureControlServer: setTemperature called setting selectedTemperatureLevel to ${request.targetTemperatureLevel}: ${this.state.supportedTemperatureLevels[request.targetTemperatureLevel]}`);
-      this.state.selectedTemperatureLevel = request.targetTemperatureLevel;
-    } else {
-      device.log.error(`MatterbridgeLevelTemperatureControlServer: setTemperature called with invalid targetTemperatureLevel ${request.targetTemperatureLevel}`);
-    }
-  }
-}
-
-export class MatterbridgeNumberTemperatureControlServer extends TemperatureControlServer.with(TemperatureControl.Feature.TemperatureNumber, TemperatureControl.Feature.TemperatureStep) {
-  override initialize() {
-    const device = this.endpoint.stateOf(MatterbridgeServer);
-    device.log.info(`MatterbridgeNumberTemperatureControlServer initialized with temperatureSetpoint ${this.state.temperatureSetpoint} minTemperature ${this.state.minTemperature} maxTemperature ${this.state.maxTemperature} step ${this.state.step}`);
-  }
-
-  override setTemperature(request: TemperatureControl.SetTemperatureRequest): MaybePromise {
-    const device = this.endpoint.stateOf(MatterbridgeServer);
-    device.log.info(`SetTemperature (endpoint ${this.endpoint.maybeId}.${this.endpoint.maybeNumber})`);
-    device.commandHandler.executeHandler('setTemperature', { request, cluster: TemperatureControlServer.id, attributes: this.state, endpoint: this.endpoint });
-    if (request.targetTemperature !== undefined && request.targetTemperature >= this.state.minTemperature && request.targetTemperature <= this.state.maxTemperature) {
-      device.log.debug(`MatterbridgeNumberTemperatureControlServer: setTemperature called setting temperatureSetpoint to ${request.targetTemperature}`);
-      this.state.temperatureSetpoint = request.targetTemperature;
-    } else {
-      device.log.error(`MatterbridgeNumberTemperatureControlServer: setTemperature called with invalid targetTemperature ${request.targetTemperature}`);
-    }
   }
 }
